state: hoist open-they-are-funding states handling into peer.c, remove state.[ch]
