+++ daemon/peer.h
@@ -283,8 +283,7 @@ void peer_unexpected_pkt(struct peer *peer, const Pkt *pkt, const char *where);
 /* Peer has an issue, breakdown and fail. */
 void peer_fail(struct peer *peer, const char *caller);
 
-/* Peer has completed open, or problem (if non-NULL). */
-void peer_open_complete(struct peer *peer, const char *problem);
+void peer_watch_anchor(struct peer *peer, int depth);
 
 struct bitcoin_tx *peer_create_close_tx(const tal_t *ctx,
 					struct peer *peer, u64 fee);
